\name{orcv-package}
\alias{orcv-package}
\alias{orcv}
\docType{package}
\title{
\packageTitle{orcv}
}
\description{
\packageDescription{orcv}
}
\details{
Asynchronous object messaging for R
}
\author{
\packageAuthor{orcv}

Maintainer: \packageMaintainer{orcv}
}
\keyword{package}

\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{orcv}}
\section{Documentation Index}{\packageIndices{orcv}}
